backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
boto3
python-dotenv